Add Langmuir calculation, refactor other calculations, adjust docs, small fixes
(rewording) Added further discussion to the ipython notebook
documentation, and added some new notebooks such as the isotherm
notebook to introduce some basics of gas adsorption.  The
documentation is not meant to be comprehensive.  The goal is to
provide just enough background to see what is going on in each
calculation, and point the reader in the right direction.

  documentation/isotherm      (new)
  documentation/tplot         (new)
  documentation/thickness
  documentation/BET
  documentation/langmuir      (new)

(new) Added the langmuir calculation following the pattern of BET

  micromeritics/langmuir  (new)

(new) Added some new reference isotherms including one which shows
absolute pressure (used by langmuir).  The new isotherms are used in
the documentation.  Renamed the variables in the restrict_isotherm()
method so that it is clear this works for both absolute and relative
pressure.

  micromeritics/isotherm_examples
  micromeritics/util

(new) Added a new method to BET to show the BET model isotherm, so
that one can check it against the experimental isotherm.  Adjusted the
BET and t-plot calculations so that they take the entire isotherm data
set. They both calculate their line fits according to the user
specified range.  Added extra information to the returned touple.

  micromeritics/bet
  micromeritics/tplot

(adjust) Adjusted the plots package to make it a little more flexible
and easier to use.  Created some re-useable string definitions.  NOTE:
this package is only for convenience and kept separate from
calculations.

  micromeritics/plots

(fix) The KrukJaroniecSayari thickness curve was using the same
constants and function as Halsey. This was fixed.

  micromeritics/thickness

(refactor) Adjusted the example scripts to reflec above changes.
Added new example scripts.  The scripts are to demonstrate usage in as
fast a way as possible. They are similar to the documentation python
notebooks.

  bet_example
  tplot_example       (new)
  bet_example
  thickness_example
  langmuir_example    (new)

(refactor) Adjusted unit test scripts.  Created a new test script for
langmuir

  micromeritics/tests/test_bet
  micromeritics/tests/test_tplot
  micromeritics/tests/test_langmuir  (new)

(refactor) Changed the constants package to be consistent with scipy.
NOTE: the value of VOLGASTP changed ever so slightly which required a
small change in the BET test script.

  micromeritics/constants

(testing) Ran all the unit test scripts to ensure no unexpected
problems with the calculations.  Also ran the ipython notebooks and
example scripts to ensure these were working as intended.